+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## Unreleased
+
+- Added new beta `RateService`, accessible via `myClient.Beta.Rate`.
+- Added `RetrieveStatelessRate` function under beta `RateService` to pull stateless rates when shipment data is provided
+- Adds `GetLowestStatelessRate` function under `RateService` and under `Calculations.Rate` to filter the lowest stateless rate
+- Adds new `GetLowest` instance functions, callable on `List<Rate>`, `List<Smartrate>` and `List<StatelessRate>`, to filter the lowest rate
+
 ## v4.3.0 (2023-01-18)
 
 - Added payload functions `RetrieveAllPayloadsForEvent` and `RetrievePayloadForEvent` methods, accessible via `myClient.Event` service.